Arudino+Particle
Uploading first versions of code for the flow meter sensor to take
readings and output them in the serial monitor for both Arduino IDE and
Particle.io cloud platform. The Particle.io code modification also has
a webhook.json file to allow for native iOS push notifications through
PushBullet. Right now it’s working through IFTTT but going to code it
in the Particle IDE for the next code deploy.